@@ -15,19 +15,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^15.0.0",
-    "@angular/animations": "^15.0.0",
+    "@angular-devkit/build-angular": "^16.2.0",
+    "@angular/animations": "^16.2.2",
     "@angular/cdk": "^14.2.7",
-    "@angular/cli": "^15.0.0",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/compiler-cli": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
+    "@angular/cli": "^16.2.0",
+    "@angular/common": "^16.2.2",
+    "@angular/compiler": "^16.2.2",
+    "@angular/compiler-cli": "^16.2.2",
+    "@angular/core": "^16.2.2",
+    "@angular/forms": "^16.2.2",
     "@angular/material": "^14.2.7",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
+    "@angular/platform-browser": "^16.2.2",
+    "@angular/platform-browser-dynamic": "^16.2.2",
+    "@angular/router": "^16.2.2",
     "@types/chart.js": "^2.9.29",
     "bootstrap": "^4.6.0",
     "chart.js": "^2.9.4",
@@ -38,8 +38,8 @@
     "quill": "^1.3.6",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
-    "typescript": "~4.8.4",
-    "zone.js": "~0.11.4"
+    "typescript": "~5.1.6",
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
     "@types/jasmine": "~4.3.0",
