@@ -1,17 +1,17 @@
-import { HttpErrorResponse } from '@angular/common/http';
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
-import { MatSnackBar } from '@angular/material/snack-bar';
+import { ChangeDetectorRef, Component, NgZone, OnDestroy, OnInit } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { MatSnackBar, } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { Observable, Subject } from 'rxjs';
-import { takeUntil, tap } from 'rxjs/operators';
-import { UserProfileDto } from '../../../models/user.model';
+import { takeUntil } from 'rxjs/operators';
 import { AuthenticationService } from '../../../services/authentication.service';
 import { UserService } from '../../../services/user.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogModel, ConfirmDialogComponent } from '../../confirm-dialog/confirm-dialog.component';
+import { failedSnackBarConfig, successSnackBarConfig } from 'src/app/models/snackbars';
 
 interface BrokerPayload {
-  controlName: string;
-  text: string;
+  brokerName: string;
   isBrokerToken: boolean;
 }
 
@@ -22,26 +22,36 @@ interface BrokerPayload {
 })
 export class SettingsComponent implements OnInit, OnDestroy {
   private destroyed$ = new Subject<void>();
+  private VALIDATION_SUCCESS = "Token is valid. Click on tab 'Create Bot'."
+  private VALIDATION_FAILED = "Token is not valid. Check the value and try again. "
 
   public hide = true;
-  public loading = true;
+  public loading = false;
+  public isDemoValue = false;
+  public checkboxTooltipText = "To use the demo token, check the box"
 
-  public profileForm: UntypedFormGroup;
-  public tinkoffTokenControl: UntypedFormControl;
-  public isTinkoffTokenControl: UntypedFormControl;
+  public profileForm: FormGroup;
+  public tinkoffTokenControl: FormControl;
+
+  public broker: BrokerPayload = {
+    brokerName: 'Tinkoff',
+    isBrokerToken: false,
+  }
 
   constructor(
-    private formBuilder: UntypedFormBuilder,
+    private formBuilder: FormBuilder,
     private router: Router,
     private userService: UserService,
     private authenticationService: AuthenticationService,
+    private detector: ChangeDetectorRef,
+    private zone: NgZone,
     private snackbar: MatSnackBar,
+    public dialog: MatDialog
   ) {
     if (!this.authenticationService.currentUserValue) {
       this.router.navigate(['/login']);
     }
     this.tinkoffTokenControl = this.formBuilder.control(null);
-    this.isTinkoffTokenControl = this.formBuilder.control(null);
     this.profileForm = this.getProfileForm();
   }
 
@@ -49,15 +59,9 @@ export class SettingsComponent implements OnInit, OnDestroy {
     return this.profileForm.valid && this.profileForm.touched;
   }
 
-  public broker: BrokerPayload = {
-    controlName: 'isTinkoffToken',
-    text: 'Tinkoff',
-    isBrokerToken: false,
-  }
-
   ngOnInit() {
-    this.getUserProfile().subscribe()
-    this.getTinkoffControlChanges().subscribe();
+    this.fetchUserProfile()
+    this.setTinkoffControlChanges()
   }
 
   ngOnDestroy(): void {
@@ -65,10 +69,38 @@ export class SettingsComponent implements OnInit, OnDestroy {
     this.destroyed$.complete();
   }
 
+  public onIsDemoCheckboxChanged(){
+    this.detector.markForCheck()
+    this.detector.detectChanges()
+  }
+
+  public onTokenValidation(): void {
+    this.userService.getValidateToken(this.isDemoValue).subscribe(
+      res => {
+        console.log("🚀 ~ file: settings.component.ts:70 ~ SettingsComponent ~ onTokenValidation ~ res:", res)
+        const message = res.isValid
+          ? this.VALIDATION_SUCCESS
+          : this.VALIDATION_FAILED
+        this.showMessage(message, res.isValid ? true : false)
+      },
+      (errorMessage: string) => {
+        this.showMessage(errorMessage, false);
+    });
+  }
+
+  /**
+   * NgZone reired for synchronize loading
+   */
   public onRemoveToken(item: BrokerPayload): void {
-    item.isBrokerToken = false;
-    this.isTinkoffTokenControl.patchValue(false);
-    this.profileForm.markAsTouched();
+    this.zone.run(() => {
+      this.confirmDialog().subscribe(res => {
+        if (res) {
+          item.isBrokerToken = false;
+          this.tinkoffTokenControl.reset();
+          this.profileForm.markAsTouched();
+        }
+      });
+    })
   }
 
   public onSubmit(): void {
@@ -77,69 +109,65 @@ export class SettingsComponent implements OnInit, OnDestroy {
     }
     this.loading = true;
 
-    const profile = this.profileForm.value as UserProfileDto;
+    const profileDto = this.profileForm.value;
     this.profileForm.disable();
-    this.userService.update(profile)
+    this.userService.update(profileDto.email, profileDto.tinkoffToken)
       .pipe(takeUntil(this.destroyed$))
-      .subscribe(_ => {
+      .subscribe(tokenDto => {
+        this.showMessage("Changes saved", true);
         this.loading = false;
-        this.showMessage("Changes saved");
         this.profileForm.enable();
         this.profileForm.markAsUntouched();
 
+        this.authenticationService.setCurrentUser(profileDto.email, tokenDto)
         this.broker.isBrokerToken = !!this.tinkoffTokenControl.value
 
-      }, (error: HttpErrorResponse) => {
+      }, (errorMessage: string) => {
         this.profileForm.enable();
         this.loading = false;
-        this.showErrorMessage(error);
+        this.showMessage(errorMessage, false);
       });
   }
 
-  private getUserProfile(): Observable<UserProfileDto> {
+  private fetchUserProfile(): void {
     this.loading = true;
-    return this.userService.get().pipe(
-      takeUntil(this.destroyed$),
-      tap(profile => {
+    this.userService.get()
+      .subscribe(profile => {
         this.loading = false;
         this.profileForm.patchValue(profile)
-        this.broker.isBrokerToken = !!profile.isTinkoffToken
+        this.broker.isBrokerToken = !!profile.tinkoffToken
+
+        this.detector.markForCheck()
+        this.detector.detectChanges()
       })
-    )
   }
 
-  private getTinkoffControlChanges(): Observable<any> {
-    return this.tinkoffTokenControl.valueChanges.pipe(
-      takeUntil(this.destroyed$),
-      tap(value => this.isTinkoffTokenControl.setValue(!!value))
-    )
+  private setTinkoffControlChanges(): void {
+    this.tinkoffTokenControl.valueChanges
+      .pipe(takeUntil(this.destroyed$)).subscribe()
   }
 
-  private getProfileForm(): UntypedFormGroup {
-    return this.formBuilder.group({
-      email: [null, [Validators.required, Validators.email]],
-      // name: [null],
-      // lastName: [null],
-      // bitrhDate: [null],
-      // location: [null],
-      // gender: [null],
-
-      tinkoffToken: this.tinkoffTokenControl,
-      isTinkoffToken: this.isTinkoffTokenControl,
+  private getProfileForm(): FormGroup {
+    return new FormGroup({
+      email: new FormControl(null, [Validators.required, Validators.email]),
+      tinkoffToken: this.tinkoffTokenControl
     });
   }
 
-  private showErrorMessage(httpError: HttpErrorResponse) {
-    if (httpError.error.array && httpError.error.array.length) {
-      const arr = httpError.error.array as Array<any>;
-      arr.forEach(el => {
-        this.snackbar.open(el.msg, 'OK', { duration: 6000 });
-      })
-    } else {
-      this.snackbar.open(httpError.error.msg, 'OK', { duration: 6000 });
-    }
+  private confirmDialog(): Observable<any> {
+    const dialogData = new ConfirmDialogModel("Token remove", "Are you sure you want to remove broker?");
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      maxWidth: "400px",
+      data: dialogData
+    });
+
+    return dialogRef.afterClosed();
   }
-  private showMessage(message: any) {
-    this.snackbar.open(message, 'OK', { duration: 3000 });
+
+  private showMessage(message: string, duration: boolean): void {
+    this.zone.run(() => { setTimeout(() => {
+      this.snackbar.open(message, 'OK',  duration ? successSnackBarConfig : failedSnackBarConfig)}, 0)
+    });
   }
 }
