@@ -1,23 +1,12 @@
-﻿import { InvestorTypeCharacter } from "../models/enums";
-import { TradingBot } from "../models/trading-bot.model";
+﻿import { TradingBot } from "./trading-bot.model";
 
 export class User {
   token: string;
   email?: string;
-  profile?: UserProfile
-  riskType?: InvestorTypeCharacter;
   bots?: TradingBot[]
 
   tinkoffToken?: string;
 }
-
-export interface UserProfile {
-  name: string;
-  lastName: string;
-  bitrhDate: Date;
-  gender: string;
-  location: string;
-};
 
 export interface UserDto {
   email: string;
@@ -26,12 +15,18 @@ export interface UserDto {
 
 export interface UserProfileDto {
   email: string;
-  name: string;
-  lastName: string;
-  bitrhDate: Date;
-  gender: string;
-  location: string;
-
   tinkoffToken?: string;
-  isTinkoffToken: boolean;
 };
+
+export class UserTokenDto {
+  token: string
+}
+
+export class UserLocalStorage {
+  email: string;
+  token: string
+}
+
+export class ValidateTokenDto {
+  isValid: boolean
+}
