@@ -31,31 +31,26 @@ import { MatRadioModule } from '@angular/material/radio';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatDividerModule } from '@angular/material/divider';
+import {MatCheckboxModule} from '@angular/material/checkbox';
 import { WidgetsComponent } from './components/_OLD/main-dashboard/widgets/widgets.component';
 import { LoginComponent } from './components/account/login/login.component';
 import { RegisterComponent } from './components/account/register/register.component';
-import { BotStatsDialogComponent } from './components/table-bots/bot-stats-dialog/bot-stats-dialog.component';
 import { TableBotsComponent } from './components/table-bots/table-bots.component';
 import { AppComponent } from './app.component';
 import { routing } from './app.routing';
 import { AboutComponent } from './components/navigation-bar/about/about.component';
 import { SettingsComponent } from './components/account/settings/settings.component';
 import { CreateBotComponent } from './components/create-bot/create-bot.component';
-import { MainDashboardComponent } from './components/_OLD/main-dashboard/main-dashboard.component';
 import { DemoModeDialog, NavigationBarComponent } from './components/navigation-bar/navigation-bar.component';
 import { JwtInterceptor } from './helpers/jwt.interceptor';
 import { ErrorInterceptor } from './helpers/error.interceptor';
-import { CreateBotInterceptor } from './components/create-bot/create.bot.interceptor';
-import { SettingsInterceptor } from './components/account/settings/settings.interceptor';
-import { DashboardInterceptor } from './components/_OLD/main-dashboard/main-dashboard.interceptor';
 import { OperationsComponent } from './components/operations/operations.component';
-import { LoginInterceptor } from './components/account/login/login.interceptor';
+import { ConfirmDialogComponent } from './components/confirm-dialog/confirm-dialog.component';
 
 @NgModule({
     declarations: [
         AppComponent,
         NavigationBarComponent,
-        MainDashboardComponent,
         SettingsComponent,
         LoginComponent,
         RegisterComponent,
@@ -63,9 +58,9 @@ import { LoginInterceptor } from './components/account/login/login.interceptor';
         CreateBotComponent,
         TableBotsComponent,
         WidgetsComponent,
-        BotStatsDialogComponent,
         OperationsComponent,
-        DemoModeDialog
+        DemoModeDialog,
+        ConfirmDialogComponent
     ],
     imports: [
         BrowserModule,
@@ -104,17 +99,18 @@ import { LoginInterceptor } from './components/account/login/login.interceptor';
         MatDatepickerModule,
         MatNativeDateModule,
         MatExpansionModule,
-        MatDividerModule
+        MatDividerModule,
+        MatCheckboxModule,
     ],
     providers: [
         MatDatepickerModule,
         MatNativeDateModule,
         { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
         { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
-        { provide: HTTP_INTERCEPTORS, useClass: CreateBotInterceptor, multi: true },
-        { provide: HTTP_INTERCEPTORS, useClass: SettingsInterceptor, multi: true },
-        { provide: HTTP_INTERCEPTORS, useClass: DashboardInterceptor, multi: true },
-        { provide: HTTP_INTERCEPTORS, useClass: LoginInterceptor, multi: true },
+        // { provide: HTTP_INTERCEPTORS, useClass: CreateBotInterceptor, multi: true },
+        // { provide: HTTP_INTERCEPTORS, useClass: SettingsInterceptor, multi: true },
+        // { provide: HTTP_INTERCEPTORS, useClass: DashboardInterceptor, multi: true },
+        // { provide: HTTP_INTERCEPTORS, useClass: LoginInterceptor, multi: true },
     ],
     bootstrap: [AppComponent]
 })
