@@ -1,8 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { HttpErrorResponse } from '@angular/common/http';
 import { AuthenticationService } from '../../../services/authentication.service';
 
 @Component({
@@ -15,10 +14,10 @@ export class LoginComponent implements OnInit {
 
   public hide = true;
   public loading = false;
-  public loginForm: UntypedFormGroup;
+  public loginForm: FormGroup;
 
   constructor(
-    private formBuilder: UntypedFormBuilder,
+    private formBuilder: FormBuilder,
     private route: ActivatedRoute,
     private router: Router,
     private authenticationService: AuthenticationService,
@@ -56,35 +55,26 @@ export class LoginComponent implements OnInit {
       .subscribe(_ => {
         this.router.navigate([this.returnUrl]);
         this.loading = false;
-      }, (error: HttpErrorResponse) => {
+      }, (errorMessage: string) => {
         this.loading = false;
-        this.showErrorMessage(error);
+        this.showMessage(errorMessage, false);
       });
   }
 
   public onForgetPassword(): void {
-    this.showMessage('Функция пока не доступна');
+    this.showMessage('Функция пока не доступна', true);
   }
 
-  private getLoginForm(): UntypedFormGroup {
+  private getLoginForm(): FormGroup {
     return this.formBuilder.group({
       email: ['', [Validators.required, Validators.email]],
       password: ['', Validators.required],
     });
   }
 
-  private showErrorMessage(httpError: HttpErrorResponse): void {
-    if (httpError.error.array && httpError.error.array.length) {
-      const arr = httpError.error.array as Array<any>;
-      arr.forEach(el => {
-        this.snackbar.open(el.msg, 'OK', { duration: 6000 });
-      })
-    } else {
-      this.snackbar.open(httpError.error.msg, 'OK', { duration: 6000 });
-    }
-  }
-
-  private showMessage(message: string): void {
-    this.snackbar.open(message, 'OK', { duration: 3000 });
+  private showMessage(message: string, duration: boolean): void {
+    duration
+    ? this.snackbar.open(message, 'OK', { duration: 5000 })
+    : this.snackbar.open(message, 'OK')
   }
 }
