@@ -2,7 +2,7 @@ import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest, HttpResponse } fr
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
 import { delay } from 'rxjs/operators';
-import { UserProfile } from '../../../models/user.model';
+import { UserProfileDto } from '../../../models/user.model';
 import { HttpUtils } from '../../../services/http-utils.service';
 import { UserService } from '../../../services/user.service';
 
@@ -16,8 +16,9 @@ export class SettingsInterceptor implements HttpInterceptor {
     private httpUtils: HttpUtils
   ) { }
 
+  // TODO add global flag isDemo
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    if (req.url.includes('/profile')) {
+    if (req.url.includes('/user')) {
       this.httpUtils.log(req);
       switch (req.method) {
         case 'GET':
@@ -34,7 +35,7 @@ export class SettingsInterceptor implements HttpInterceptor {
     return next.handle(req);
   }
 
-  private getBody(): UserProfile {
+  private getBody(): UserProfileDto {
     return this.service.mockUserProfile;
   }
 
