@@ -1,10 +1,11 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, NgZone, OnInit } from '@angular/core';
 import { tap } from 'rxjs/operators';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTableDataSource } from '@angular/material/table';
-import { TradingBot } from '../../models/trading-bot.model';
+import { CronStatus, TradingBot } from '../../models/trading-bot.model';
 import { TradingBotsService } from '../../services/trading-bots.service';
+import { successSnackBarConfig, failedSnackBarConfig } from 'src/app/models/snackbars';
 
 @Component({
   selector: 'app-table-bots',
@@ -12,36 +13,41 @@ import { TradingBotsService } from '../../services/trading-bots.service';
   styleUrls: ['./table-bots.component.scss'],
 })
 export class TableBotsComponent implements OnInit {
-  displayedColumns: string[] = ['id', 'ticker', 'strategy', 'timeframe', 'profit', 'status', 'actions'];
+  // displayedColumns: string[] = ['id', 'ticker', 'strategy', 'timeframe', 'profit', 'status', 'actions'];
+  displayedColumns: string[] = ['id', 'ticker', 'status', 'actions'];
   dataSource: MatTableDataSource<TradingBot> = new MatTableDataSource();
   isLoading = true;
 
   readonly ONE_DAY=1000*60*60;
 
-  constructor(private readonly tradingBotsService: TradingBotsService, private readonly snackBar: MatSnackBar, public dialog: MatDialog) { }
+  constructor(
+    private readonly tradingBotsService: TradingBotsService,
+    private readonly snackBar: MatSnackBar,
+    public dialog: MatDialog,
+    private zone: NgZone,
+  ) { }
 
   ngOnInit() {
     this.fetchBots();
   }
 
   public stopBot(bot: TradingBot): void {
-    bot.isActive = !bot.isActive;
+    bot.status = bot.status === CronStatus.Scheduled ? CronStatus.Stopped : CronStatus.Scheduled;
     this.tradingBotsService
       .updateBotData(bot)
-      .pipe(
-        tap(
-          _ => {
-            (bot.isActive === false)
-              ? this.showMessage('Request to stop bot sent')
-              : this.showMessage('Request to start bot sent')
-          },
-          err => this.showMessage(err)
-        ))
-      .subscribe();
+      .subscribe(res => {
+        (res.status === CronStatus.Scheduled)
+          ? this.showMessage('Request to stop bot sent', true)
+          : this.showMessage('Request to start bot sent', true)
+      },
+        err => this.showMessage(err, false)
+      );
   }
 
   public delete(bot: TradingBot): void {
-    this.tradingBotsService.deleteBotData(bot).subscribe();
+    this.tradingBotsService.deleteBotData(bot).subscribe(
+      () => { this.showMessage(`Bot with ticker: ${bot.ticker} was removed.`, true )},
+      err => this.showMessage(err, false));
     this.fetchBots();
   }
 
@@ -49,16 +55,19 @@ export class TableBotsComponent implements OnInit {
     this.tradingBotsService.getUserBots().subscribe(bots => {
       this.dataSource.data = bots;
 
-      bots.forEach(bot => {
-        bot.workedTime = msToTime(Date.now() - bot.createdDate)
-      })
+      // bots.forEach(bot => {
+      //   bot.workedTime = msToTime(Date.now() - bot.createdDate)
+      // })
 
       this.isLoading = false;
     });
   }
 
-  private showMessage(msg: any) {
-    this.snackBar.open(msg, undefined, { duration: 2000 });
+  private showMessage(message: string, duration: boolean): void {
+    this.zone.run(() => { setTimeout(() => {
+        this.snackBar.open(message, 'OK',  duration ? successSnackBarConfig : failedSnackBarConfig)
+      }, 0)
+    });
   }
 
   // public openDialog(row: TradingBot): void {
@@ -77,12 +86,12 @@ export class TableBotsComponent implements OnInit {
   // }
 }
 
-const msToTime = (duration: number) => {
-  let minutes = Math.floor((duration / (1000 * 60)) % 60)
-  let hours = Math.floor((duration / (1000 * 60 * 60)) % 24)
+// const msToTime = (duration: number) => {
+//   let minutes = Math.floor((duration / (1000 * 60)) % 60)
+//   let hours = Math.floor((duration / (1000 * 60 * 60)) % 24)
 
-  const hoursStr = (hours < 10) ? "0" + hours : hours.toString();
-  const minutesStr = (minutes < 10) ? "0" + minutes : minutes.toString();
+//   const hoursStr = (hours < 10) ? "0" + hours : hours.toString();
+//   const minutesStr = (minutes < 10) ? "0" + minutes : minutes.toString();
 
-  return hoursStr + "h " + minutesStr + "m "
-}
+//   return hoursStr + "h " + minutesStr + "m "
+// }
