@@ -18,7 +18,7 @@ export class ErrorInterceptor implements HttpInterceptor {
           location.reload();
         }
 
-        const error = err.error.message || err.statusText;
+        const error = err.error;
         return throwError(error);
       })
     );
